search books can show correct status and fix no cover and no author
REQUIRED
When a book in results doesn't have a thumbnail and is missing the
imageLinks property, then no results display for that search term. Try
searching for 'biography' to test this as the book "My First Life" by
Brenda Dean Paul that is normally in the results does not have the
imageLinks property.

The best solution is probably to create a default cover image to be
displayed if a book doesn't have a cover.

You could test this out using this something like this as a default
cover http://via.placeholder.com/128x193?text=No%20Cover (Use the URL in
place of an image. You can customize the width, height, and text by
changing the URL.)

SUGGESTION
If search results contain a book without any authors, then no results
are shown.

You'll need to conditionally render the statement with book.authors if
you use join by checking that book has the authors property first as
some books do not have an author. (Example: search for 'poetry' and the
book "Principia Latina, part iii. An introduction to Latin poetry" does
not have an author)

Here's one way you could use conditional rendering to resolve this:

<div className="book-authors">{book.authors ? book.authors.join(', ') :
''}</div>

REQUIRED
Books in search results should have the shelf they're on selected in the
dropdown for that book, and if a book is not on a shelf then it should
not have any shelf (or a "None" shelf) selected.

Please note that the books in search results you get from
BooksAPI.search() don't have any shelf data, so you'll have to add the
correct shelf after comparing the books in search results with from the
data you obtained from BooksAPI.getAll() -- it would be best if you
could pass this data down as props instead of calling BooksAPI.getAll()
again in this component.

Example

Here I have added the book 'Pro React' to the 'Want to Read' shelf, and
it shows up under that shelf on the main page.

I then click the link to go to the search page and enter 'react' into
the search input. In the results, I click the dropdown for the book 'Pro
React' and it does not have the 'Want to Read' shelf selected like it
should.